move BaseCLI
