+++ sup3r/pipeline/forward_pass_cli.py
@@ -11,7 +11,7 @@ import os
 from sup3r.utilities import ModuleName
 from sup3r.version import __version__
 from sup3r.pipeline.forward_pass import ForwardPassStrategy, ForwardPass
-from sup3r.cli.base import BaseCLI
+from sup3r.utilities.cli_utils import BaseCLI
 
 
 logger = logging.getLogger(__name__)
