+++ sup3r/postprocessing/data_collect_cli.py
@@ -9,7 +9,7 @@ import copy
 from sup3r.utilities import ModuleName
 from sup3r.version import __version__
 from sup3r.postprocessing.collection import Collector
-from sup3r.cli.base import BaseCLI
+from sup3r.utilities.cli_utils import BaseCLI
 
 
 logger = logging.getLogger(__name__)
