+++ sup3r/preprocessing/data_extract_cli.py
@@ -8,7 +8,7 @@ import logging
 import sup3r
 from sup3r.utilities import ModuleName
 from sup3r.version import __version__
-from sup3r.cli.base import BaseCLI
+from sup3r.utilities.cli_utils import BaseCLI
 
 
 logger = logging.getLogger(__name__)
