+++ sup3r/qa/qa_cli.py
@@ -8,7 +8,7 @@ import logging
 from sup3r.utilities import ModuleName
 from sup3r.version import __version__
 from sup3r.qa.qa import Sup3rQa
-from sup3r.cli.base import BaseCLI
+from sup3r.utilities.cli_utils import BaseCLI
 
 
 logger = logging.getLogger(__name__)
